search_engine.py
import json
from openai_api import query_openai

# Load the configuration file for search engine settings
with open('config.json') as config_file:
    config = json.load(config_file)

SEARCH_ENGINE_SETTINGS = config['search_engine_settings']

def search(query, language=SEARCH_ENGINE_SETTINGS['default_search_language'], result_limit=SEARCH_ENGINE_SETTINGS['result_limit']):
    """
    This function uses the OpenAI API to perform a search based on the user's query.

    :param query: The user's search query.
    :param language: The language for the search results.
    :param result_limit: The maximum number of search results to return.
    :return: A list of search results.
    """
    # Modify the query to specify the language and request a certain number of results
    modified_query = f"Search for articles in {language} about: {query}. Provide a maximum of {result_limit} results."

    # Use the OpenAI API to get a response for the modified query
    response = query_openai(modified_query)

    # Process the response to extract search results
    if response:
        # Here you would implement the logic to parse the response into a list of search results.
        # For the sake of this example, we'll assume the response is a plain text list of URLs or titles.
        # In a real-world scenario, you would likely need to do more complex parsing, possibly involving
        # additional API calls to fetch search results from a database or a web search API.
        search_results = response.split('\n')[:result_limit]
        return search_results
    else:
        return []

# Example usage:
# results = search("machine learning")
# for result in results:
#     print(result)
